--- src/main/java/io/carbonchain/hiring/java/Controller.java
@@ -1,4 +1,0 @@
-package io.carbonchain.hiring.java;
-
-public interface Controller {
-}
