--- src/main/java/io/carbonchain/hiring/java/Middleware.java
@@ -1,5 +1,0 @@
-package io.carbonchain.hiring.java;
-
-public interface Middleware {
-  public Request handle(Request request);
-}
