+++ src/main/java/io/carbonchain/hiring/java/app/Core.java
@@ -1,4 +1,9 @@
-package io.carbonchain.hiring.java;
+package io.carbonchain.hiring.java.app;
+
+import io.carbonchain.hiring.java.Request;
+import io.carbonchain.hiring.java.controller.Controller;
+import io.carbonchain.hiring.java.exception.ApplicationException;
+import io.carbonchain.hiring.java.middleware.Middleware;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.HashMap;
