+++ src/main/java/io/carbonchain/hiring/java/controller/Controller.java
@@ -1,0 +1,4 @@
+package io.carbonchain.hiring.java.controller;
+
+public interface Controller {
+}
