+++ src/main/java/io/carbonchain/hiring/java/middleware/Middleware.java
@@ -1,0 +1,7 @@
+package io.carbonchain.hiring.java.middleware;
+
+import io.carbonchain.hiring.java.Request;
+
+public interface Middleware {
+  public Request handle(Request request);
+}
